Quickfix: make tests importable without mock
Sublime thinks the tests are a part of the plugin and tries
to run them. This failed since they import "mock".
It's now possible to import the tests without "mock" installed,
but running them will fail at runtime then. That's okay since
sublime does not try to run tests.

A more elegant solution would be telling package control that the
tests module is not part of the plugin. I don't know if that is
possible.