Avoid creating temporary files; use stdin instead
Racer supports a 'substitute file' as its last argument.
When provided, the primary path is still used to find adjacent
modules and is named in the MATCH lines (if relevant),
but the actual content is read from the substitute path.

Conveniently, the substitute file can be stdin (via '-').

With this change, the current view contents are piped over stdin
via '-' (instead of creating a temporary file). This avoids transiently
seeing temp files in the folders view, accidentally adding them to git,
etc.